DOCKER-929 Support the docker v2.2 manifest format Reviewed by: Trent Mick <trentm@gmail.com> Approved by: Trent Mick <trentm@gmail.com>

+++ test/lib/cli.js
@@ -79,14 +79,17 @@ function cliInspect(t, opts, callback) {
         var obj;
         var pe;
 
-        t.ifErr(err, 'docker inspect ' + opts.id);
-        t.equal(stderr, '', 'stderr should be empty');
-
-        // XXX: allow setting opts.expectedErr
         if (err) {
-            common.done(t, callback, err);
+            if (opts.expectedErr) {
+                common.expCliErr(t, stderr, opts.expectedErr, callback);
+            } else {
+                t.ifErr(err, 'docker inspect ' + opts.id);
+                common.done(t, callback, err);
+            }
             return;
         }
+
+        t.equal(stderr, '', 'stderr should be empty');
 
         if (!stdout) {
             var stdoutErr = new Error('no stdout!');
@@ -293,6 +296,36 @@ function cliRmAllCreated(t) {
 
 
 /**
+ * Removes all docker VMs that have the given name prefix.
+ */
+function cliRmContainersWithNamePrefix(tt, prefix) {
+    tt.test('remove old containers', function (t) {
+        cliPs(t, {args: '-a'}, function (err, entries) {
+            t.ifErr(err, 'docker ps');
+
+            var containersToRemove = entries.filter(function (entry) {
+                return (entry.names.substr(0, prefix.length) === prefix);
+            });
+
+            vasync.forEachParallel({
+                inputs: containersToRemove,
+                func: function _delOne(entry, cb) {
+                    cliRm(t, {args: '-f ' + entry.container_id}, onRemove);
+                    function onRemove(err2) {
+                        t.ifErr(err2, 'rm container ' + entry.container_id);
+                        cb(err2);
+                    }
+                }
+            }, function (forEachErr) {
+                tt.ifErr(forEachErr);
+                t.end();
+            });
+        });
+    });
+}
+
+
+/**
  * `docker create <cmd>`
  */
 function cliCreate(t, opts, callback) {
@@ -460,7 +493,7 @@ function cliRm(t, opts, callback) {
         t.ifErr(err, 'docker rm ' + opts.args);
         t.equal(stderr, '', 'stderr');
 
-        callback(err);
+        common.done(t, callback, err);
     });
 }
 
@@ -474,9 +507,14 @@ function cliRmi(t, opts, callback) {
     assert.string(opts.args, 'opts.args');
 
     ALICE.docker('rmi ' + opts.args, function (err, stdout, stderr) {
+        if (opts.expectedErr) {
+            common.expCliErr(t, stderr, opts.expectedErr, callback);
+            return;
+        }
+
         t.ifErr(err, 'docker rmi ' + opts.args);
-        t.equal(stderr, '', 'stderr');
-        callback(err);
+        t.equal(stderr, '', 'docker rmi stderr should be empty');
+        common.done(t, callback, err);
     });
 }
 
@@ -591,6 +629,7 @@ module.exports = {
     rm: cliRm,
     rmi: cliRmi,
     rmAllCreated: cliRmAllCreated,
+    rmContainersWithNamePrefix: cliRmContainersWithNamePrefix,
     run: cliRun,
     stop: cliStop,
     start: cliStart,
